style(msvc): fix C4244/C4267 in ActorMultiVertex/BitmapText/RageFileBasic/ScreenDebugOverlay (item 2)
All static_cast, documenting existing intentional narrowing -- no
logic changes. sm_tests unchanged at 5966/226.

- ActorMultiVertex.cpp: Lua bindings pushing size_t (vertex/state
  counts and indices) into lua_pushnumber's lua_Number / lua_rawgeti's
  int; a size_t modulo/subtraction into int locals; VertexIndex
  (size_t) into SetVertexPos/Coords/Color(int index, ...); a size_t
  return cast into SetState(int)/AddVertices(int).
- BitmapText.cpp: std::fmin(1, floatExpr) promoting the whole call to
  double via the bare int literal -- fixed with a 1.0f literal instead
  of a cast, so the macro stays entirely float; size_t counts into int
  locals; rnd()%N (unsigned int) into RageVector3's float ctor args;
  lua_tointeger() into Attribute::length (int).
- RageFileBasic.cpp: pointer-diffs into int (buffer offset,
  FillReadBuf's available-space calc); a Write/Read size_t byte-count
  parameter added into or divided against int members/locals across
  EmptyWriteBuf/Write/the 3-arg Write overload.
- ScreenDebugOverlay.cpp: const_iterator-begin() pointer-diffs into
  int page/subscriber indices; a double (RageTimer::GetTimeSinceStart)
  into SecondsToMMSSMsMsMs's float param.

+++ src/ScreenDebugOverlay.cpp
@@ -270,7 +270,7 @@ void ScreenDebugOverlay::Init()
 	auto start = m_asPages.begin();
 	for (std::vector<RString>::const_iterator s = m_asPages.begin(); s != m_asPages.end(); ++s)
 	{
-		int iPage = s - start;
+		int iPage = static_cast<int>(s - start);
 
 		DeviceInput di;
 		bool b = GetKeyFromMap( g_Mappings.pageButton, iPage, di );
@@ -365,7 +365,7 @@ void ScreenDebugOverlay::UpdateText()
 	auto start = m_asPages.begin();
 	for (std::vector<RString>::const_iterator s = m_asPages.begin(); s != m_asPages.end(); ++s)
 	{
-		int iPage = s - start;
+		int iPage = static_cast<int>(s - start);
 		m_vptextPages[iPage]->PlayCommand( (iPage == m_iCurrentPage) ? "GainFocus" :  "LoseFocus" );
 	}
 
@@ -376,7 +376,7 @@ void ScreenDebugOverlay::UpdateText()
 	{
 		RString sPageName = (*p)->GetPageName();
 
-		int i = p - subStart;
+		int i = static_cast<int>(p - subStart);
 
 		float fY = LINE_START_Y + iOffset * LINE_SPACING;
 
@@ -477,7 +477,7 @@ bool ScreenDebugOverlay::Input( const InputEventPlus &input )
 	{
 		RString sPageName = (*p)->GetPageName();
 
-		int i = p - start;
+		int i = static_cast<int>(p - start);
 
 		// Gameplay buttons are available only in gameplay. Non-gameplay buttons
 		// are only available when the screen is displayed.
@@ -1319,7 +1319,7 @@ class DebugLineForceCrash : public IDebugLine
 class DebugLineUptime : public IDebugLine
 {
 	RString GetDisplayTitle() override { return UPTIME.GetValue(); }
-	RString GetDisplayValue() override { return SecondsToMMSSMsMsMs(RageTimer::GetTimeSinceStart()); }
+	RString GetDisplayValue() override { return SecondsToMMSSMsMsMs(static_cast<float>(RageTimer::GetTimeSinceStart())); }
 	bool IsEnabled() override { return false; }
 	void DoAndLog( RString &/* sMessageOut */ ) override {}
 };
